Remove Elasticsearch6Index and make Elasticsearch7Index the base implementation
