+++ wagtail/search/backends/elasticsearch7.py
@@ -19,7 +19,6 @@ from wagtail.search.backends.base import (
     get_model_root,
 )
 from wagtail.search.backends.elasticsearch6 import (
-    Elasticsearch6Index,
     Elasticsearch6Mapping,
     Field,
 )
@@ -36,7 +35,13 @@ class Elasticsearch7Mapping(Elasticsearch6Mapping):
         return mapping[self.get_document_type()]
 
 
-class Elasticsearch7Index(Elasticsearch6Index):
+class Elasticsearch7Index:
+    def __init__(self, backend, name):
+        self.backend = backend
+        self.es = backend.es
+        self.mapping_class = backend.mapping_class
+        self.name = name
+
     if use_new_elasticsearch_api:
 
         def put(self):
@@ -50,6 +55,47 @@ class Elasticsearch7Index(Elasticsearch6Index):
 
         def refresh(self):
             self.es.indices.refresh(index=self.name)
+
+    else:
+
+        def put(self):
+            self.es.indices.create(self.name, self.backend.settings)
+
+        def delete(self):
+            try:
+                self.es.indices.delete(self.name)
+            except NotFoundError:
+                pass
+
+        def refresh(self):
+            self.es.indices.refresh(self.name)
+
+    def exists(self):
+        return self.es.indices.exists(self.name)
+
+    def is_alias(self):
+        return self.es.indices.exists_alias(name=self.name)
+
+    def aliased_indices(self):
+        """
+        If this index object represents an alias (which appear the same in the
+        Elasticsearch API), this method can be used to fetch the list of indices
+        the alias points to.
+
+        Use the is_alias method if you need to find out if this an alias. This
+        returns an empty list if called on an index.
+        """
+        return [
+            self.backend.index_class(self.backend, index_name)
+            for index_name in self.es.indices.get_alias(name=self.name).keys()
+        ]
+
+    def put_alias(self, name):
+        """
+        Creates a new alias to this index. If the alias already exists it will
+        be repointed to this index.
+        """
+        self.es.indices.put_alias(name=name, index=self.name)
 
     def add_model(self, model):
         # Get mapping
@@ -120,6 +166,13 @@ class Elasticsearch7Index(Elasticsearch6Index):
             self.es.delete(index=self.name, id=mapping.get_document_id(item))
         except NotFoundError:
             pass  # Document doesn't exist, ignore this exception
+
+    def reset(self):
+        # Delete old index
+        self.delete()
+
+        # Create new index
+        self.put()
 
 
 class Elasticsearch7SearchQueryCompiler(BaseSearchQueryCompiler):
